+++ src/main/kotlin/com/example/kbocombo/crawler/application/PlayerClient.kt
@@ -1,0 +1,8 @@
+package com.example.kbocombo.crawler.application
+
+import com.example.kbocombo.domain.player.Player
+
+interface PlayerClient {
+
+    fun getRecentPlayers() : List<Player>
+}
