+++ src/main/kotlin/com/example/kbocombo/crawler/config/KboHttpClientConfig.kt
@@ -1,6 +1,6 @@
 package com.example.kbocombo.crawler.config
 
-import com.example.kbocombo.crawler.infrastructure.KboHttpClient
+import com.example.kbocombo.crawler.infra.KboHttpClient
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.http.HttpHeaders
