+++ src/main/kotlin/com/example/kbocombo/crawler/infra/KboHttpClient.kt
@@ -1,4 +1,4 @@
-package com.example.kbocombo.crawler.infrastructure
+package com.example.kbocombo.crawler.infra
 
 import org.springframework.http.ResponseEntity
 import org.springframework.util.MultiValueMap
