+++ src/main/kotlin/com/example/kbocombo/crawler/infra/KboPlayerDetailPageParser.kt
@@ -1,4 +1,4 @@
-package com.example.kbocombo.crawler.infrastructure
+package com.example.kbocombo.crawler.infra
 
 import com.example.kbocombo.crawler.utils.toHittingHand
 import com.example.kbocombo.crawler.utils.toPitchingHand
