--- src/main/kotlin/com/example/kbocombo/crawler/service/PlayerClient.kt
@@ -1,8 +1,0 @@
-package com.example.kbocombo.crawler.service
-
-import com.example.kbocombo.domain.player.Player
-
-interface PlayerClient {
-
-    fun findAllPlayers() : List<Player>
-}
